+++ vite.config.ts
@@ -1,11 +1,22 @@
 import { defineConfig } from "vite";
 import { resolve } from "path";
 // vite.config.ts
-import Icons from 'unplugin-icons/vite'
-import IconsResolver from 'unplugin-icons/resolver'
-import Components from 'unplugin-vue-components/vite'
+import Icons from "unplugin-icons/vite";
+import IconsResolver from "unplugin-icons/resolver";
+// import AntdvResolver from "antdv-component-resolver";
+import Components from "unplugin-vue-components/vite";
 // 引入Unocss
-import Unocss from "unocss/vite";
+// import Unocss from "unocss/vite";
+
+// import {theme} from "ant-design-vue"
+
+// const { theme } = require("ant-design-vue/lib");
+// const convertLegacyToken = require("ant-design-vue/lib/theme/convertLegacyToken");
+
+// const { defaultAlgorithm, defaultSeed } = theme;
+
+// const mapToken = defaultAlgorithm(defaultSeed);
+// const v3Token = convertLegacyToken(mapToken);
 
 function pathResolve(dir: string) {
   return resolve(__dirname, ".", dir);
@@ -15,14 +26,20 @@ function pathResolve(dir: string) {
 export default defineConfig({
   plugins: [
     Components({
-      resolvers: [
-        IconsResolver(),
-      ]
+      resolvers: [IconsResolver()], //, AntdvResolver()
     }),
-    Unocss(),
-    Icons({ autoInstall: true })
+    // Unocss(),
+    Icons({ autoInstall: true }),
   ],
   ssr: {
     noExternal: ["vitepress-plugin-nprogress"],
   },
+  // css: {
+  //   preprocessorOptions: {
+  //     less: {
+  //       math: "always",
+  //       modifyVars: theme.useToken().hashId,
+  //     },
+  //   },
+  // },
 });