+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## <small>0.0.4 (2023-07-17)</small>
+
+* feat(custom): 完善博客页面和关于页面 ([fdd5054](https://github.com/changweihua/changweihua.github.io/commit/fdd5054))
+* fix(custom): 修复样式失效 ([15ef1a7](https://github.com/changweihua/changweihua.github.io/commit/15ef1a7))
+* fix(custom): 修复unocss打包异常的问题 ([1e5740f](https://github.com/changweihua/changweihua.github.io/commit/1e5740f))
+
+
+
 ## <small>0.0.4 (2023-07-04)</small>
 
 * fix(custom): 修复样式失效 ([15ef1a7](https://github.com/changweihua/changweihua.github.io/commit/15ef1a7))